downloadAndUnzip now checks if zip file already exists to avoid superfluous downloads

+++ src/bin/tools/downloadAndUnzip.ts
@@ -2,7 +2,7 @@ import { basename as pathBasename, join as pathJoin } from "path";
 import { execSync } from "child_process";
 import fs from "fs";
 import { transformCodebase } from "./transformCodebase";
-import { rm_rf, rm, rm_r } from "./rm";
+import { rm_rf } from "./rm";
 
 /** assert url ends with .zip */
 export function downloadAndUnzip(params: { url: string; destDirPath: string; pathOfDirToExtractInArchive?: string }) {
@@ -11,22 +11,20 @@ export function downloadAndUnzip(params: { url: string; destDirPath: string; pat
     const tmpDirPath = pathJoin(destDirPath, "..", "tmp_xxKdOxnEdx");
     const zipFilePath = pathBasename(url);
 
-    rm_rf(tmpDirPath);
+    if (!fs.existsSync(pathJoin(tmpDirPath, zipFilePath))) {
+        rm_rf(tmpDirPath);
 
-    fs.mkdirSync(tmpDirPath, { "recursive": true });
-
-    execSync(`curl -L ${url} -o ${zipFilePath}`, { "cwd": tmpDirPath });
+        fs.mkdirSync(tmpDirPath, { "recursive": true });
+        
+        execSync(`curl -L ${url} -o ${zipFilePath}`, { "cwd": tmpDirPath });
+    }
 
     execSync(`unzip -o ${zipFilePath}${pathOfDirToExtractInArchive === undefined ? "" : ` "${pathOfDirToExtractInArchive}/**/*"`}`, {
         "cwd": tmpDirPath,
     });
 
-    rm(pathBasename(url), { "cwd": tmpDirPath });
-
     transformCodebase({
         "srcDirPath": pathOfDirToExtractInArchive === undefined ? tmpDirPath : pathJoin(tmpDirPath, pathOfDirToExtractInArchive),
         destDirPath,
     });
-
-    rm_r(tmpDirPath);
 }
